removed debug statements from symlink_Code.py, added autodetection for fortran compiler and disabled compiling on default

+++ code/setup/symlink_Code.py
@@ -13,17 +13,11 @@ baseDir = damaskEnv.relPath('code/')
 
 for dir in bin_link:
   for file in bin_link[dir]:
-    print baseDir
-    print dir
-    print file
     src = os.path.abspath(os.path.join(baseDir,dir,file))
-    print src
-    os.path.exists(src)
     if os.path.exists(src): 
       sym_link = os.path.abspath(os.path.join(damaskEnv.binDir(),\
                                               {True: dir,
                                                False:os.path.splitext(file)[0]}[file == '']))
-      if os.path.lexists(sym_link):
-        os.remove(sym_link)
+      if os.path.lexists(sym_link): os.remove(sym_link)
       os.symlink(src,sym_link)
       print sym_link,'-->',src
